Split mixed-tier test files so module-level pytestmark is well-defined
`test_stoichiometry.py` and `test_invariants.py` carried classes from
two pytest tiers: the former mixed unit-tier stoichiometry checks with
an integration-tier end-to-end mass-conservation class; the latter
mixed smoke-tier solver invariants with unit-tier solubility-law
invariants. Neither file could acquire a module-level `pytestmark`
because pytest stacks module and class markers: a module-level
`pytestmark = pytest.mark.unit` would pull the integration class into
the PR gate's unit selection, and vice versa.

The fix is to split each file per tier. The integration class moves
to a new `test_stoichiometry_integration.py` (`pytestmark =
[pytest.mark.integration, pytest.mark.timeout(300)]`). The four
unit-tier classes move to a new `test_invariants_unit.py` (`pytestmark
= [pytest.mark.unit, pytest.mark.timeout(30)]`). Both source files now
acquire a single tier marker at module level and drop their now-
redundant `@pytest.mark.<tier>` class decorators.

The split preserves the tier semantics exactly: the PR-gate unit
collection still finds 195 tests; smoke 107; integration 5; slow 11.
All 63 tests across the affected files pass.

The baseline ratchet clears the `missing_module_pytestmark: 2` entry
that has been the last open architectural carve-out. Total drops
46 -> 43.
