+++ tests/test_invariants.py
@@ -1,4 +1,5 @@
-"""Physics and chemistry invariants for the CALLIOPE outgassing solver.
+"""Smoke-tier physics and chemistry invariants for the CALLIOPE
+outgassing solver.
 
 Each invariant in this file asserts a property that must hold for any
 valid CALLIOPE result, independent of the specific (T_magma, fO2_shift,
@@ -7,7 +8,7 @@ each parametric sweep covers an edge of the physical regime (extreme
 reducing, extreme oxidising, low T, high T), and each class includes
 at least one sad-path test that exercises an unphysical input.
 
-The 11 invariants exercised here are:
+The 8 smoke-tier invariants exercised here are:
 
 1.  Per-element mass conservation: atm + liquid == total.
 2.  Pressure positivity: every species partial pressure is non-negative.
@@ -19,10 +20,6 @@ The 11 invariants exercised here are:
 7.  Modified equilibrium constant identity: p_B / p_A matches
     ``ModifiedKeq.<method>(T, fO2_shift)`` at the converged state.
 8.  Dissolved-mass non-negativity: every <species>_kg_liquid is >= 0.
-9.  CO2 stoichiometric atom counting: the C contribution from CO2_kg_atm
-    equals (12/44) * CO2_kg_atm at known partial pressure.
-10. Monotonicity of dissolved S vs reducing fO2 at fixed p_S2.
-11. Monotonicity of dissolved N vs reducing fO2 at fixed p_N2.
 
 Plus three Tim-flagged additions:
 
@@ -34,6 +31,13 @@ Plus three Tim-flagged additions:
     begins and the law extrapolates).
 -   p_guess warm-start verification: a warm start with the cold-solve
     result drops the restart count to zero or one.
+
+The unit-tier invariants (CO2 stoichiometric atom counting, S2 Gaillard
+monotonicity, N2 Dasgupta monotonicity, Libourel redox-independence,
+Dasgupta law at the reducing edge) live in `test_invariants_unit.py`.
+That split is needed because pytest stacks module-level and class-level
+markers; a single module-level pytestmark on a mixed-tier file would
+pull unit tests into the smoke gate.
 """
 
 from __future__ import annotations
@@ -44,14 +48,14 @@ import math
 import pytest
 
 from calliope.chemistry import ModifiedKeq
-from calliope.constants import element_list, molar_mass, volatile_species
+from calliope.constants import element_list, volatile_species
 from calliope.oxygen_fugacity import OxygenFugacity
-from calliope.solubility import SolubilityN2, SolubilityS2
 from calliope.solve import (
-    _atmosphere_mass,
     equilibrium_atmosphere,
     equilibrium_atmosphere_authoritative_O,
 )
+
+pytestmark = [pytest.mark.smoke, pytest.mark.timeout(60)]
 
 logging.getLogger('calliope').setLevel(logging.WARNING)
 
@@ -129,7 +133,6 @@ _DEFAULT_TFO2 = [
 # ===========================================================================
 
 
-@pytest.mark.smoke
 class TestMassConservationPerElement:
     """Per-element atmospheric + dissolved == total."""
 
@@ -189,7 +192,6 @@ class TestMassConservationPerElement:
 # ===========================================================================
 
 
-@pytest.mark.smoke
 class TestPressurePositivity:
     """Every species partial pressure is >= 0 at the converged solution."""
 
@@ -234,7 +236,6 @@ class TestPressurePositivity:
 # ===========================================================================
 
 
-@pytest.mark.smoke
 class TestVMRClosure:
     """sum(volume mixing ratios) == 1.0 over all volatile species."""
 
@@ -276,7 +277,6 @@ class TestVMRClosure:
 # ===========================================================================
 
 
-@pytest.mark.smoke
 class TestTotalPressureConsistency:
     """P_surf == sum of species partial pressures."""
 
@@ -310,7 +310,6 @@ class TestTotalPressureConsistency:
 # ===========================================================================
 
 
-@pytest.mark.smoke
 class TestAtmosphericMassConsistency:
     """M_atm == sum of <species>_kg_atm."""
 
@@ -334,7 +333,6 @@ class TestAtmosphericMassConsistency:
 # ===========================================================================
 
 
-@pytest.mark.smoke
 class TestFO2Reconstruction:
     """log10(p_O2 / fO2_IW_buffer(T)) == fO2_shift_derived."""
 
@@ -367,7 +365,6 @@ class TestFO2Reconstruction:
 # ===========================================================================
 
 
-@pytest.mark.smoke
 class TestModifiedKeqIdentity:
     """p_B / p_A == ModifiedKeq.<method>(T, fO2_shift) for each couple."""
 
@@ -404,7 +401,6 @@ class TestModifiedKeqIdentity:
 # ===========================================================================
 
 
-@pytest.mark.smoke
 class TestDissolvedMassNonNegativity:
     """<species>_kg_liquid is >= 0 for every species at convergence."""
 
@@ -431,157 +427,10 @@ class TestDissolvedMassNonNegativity:
 
 
 # ===========================================================================
-# Invariant 9: CO2 stoichiometric atom counting (unit-tier, no solver)
-# ===========================================================================
-
-
-@pytest.mark.unit
-class TestCO2AtomCounting:
-    """At known p_CO2, the C contribution from CO2 in the atmosphere
-    is exactly (12/44) * CO2_kg_atm via stoichiometric atom counting."""
-
-    @pytest.mark.parametrize('p_CO2_bar', [0.1, 1.0, 10.0, 100.0])
-    def test_C_atom_count_from_CO2_only(self, p_CO2_bar):
-        # Single-species atmosphere with only CO2 to isolate C contribution
-        ddict = _ddict(T=1800.0, Phi=0.0, dIW=0.0)
-        for sp in volatile_species:
-            ddict[f'{sp}_included'] = 0
-        ddict['CO2_included'] = 1
-        # Set all pressures to zero, then CO2
-        p_d = {s: 0.0 for s in volatile_species}
-        p_d['CO2'] = p_CO2_bar
-        mass_atm = _atmosphere_mass(p_d, 0.0, ddict)
-        # Atomic-C contribution expected from stoichiometry: 12.011 / 44.01
-        # times the CO2 column mass
-        expected_C_kg = mass_atm['CO2'] * molar_mass['C'] / molar_mass['CO2']
-        assert mass_atm['C'] == pytest.approx(expected_C_kg, rel=1e-12)
-
-        # Discrimination guard: the wrong stoichiometry (factor 2 for C in
-        # CO2, i.e. treating CO2 as having 2 C atoms) would double the C
-        # tally, well outside the rel=1e-12 tolerance.
-        wrong_C_kg = 2.0 * mass_atm['CO2'] * molar_mass['C'] / molar_mass['CO2']
-        assert abs(mass_atm['C'] - wrong_C_kg) > 0.5 * expected_C_kg
-
-    def test_zero_CO2_pressure_gives_zero_C(self):
-        """Sad-path: at p_CO2 = 0 the C tally from CO2 channel is zero."""
-        ddict = _ddict(T=1800.0, Phi=0.0, dIW=0.0)
-        for sp in volatile_species:
-            ddict[f'{sp}_included'] = 0
-        ddict['CO2_included'] = 1
-        p_d = {s: 0.0 for s in volatile_species}
-        mass_atm = _atmosphere_mass(p_d, 0.0, ddict)
-        assert mass_atm.get('C', 0.0) == pytest.approx(0.0, abs=1e-30)
-
-        # Discrimination guard: confirm the CO2 column mass is also zero.
-        # A stub that returned zero only for the C key (but nonzero CO2
-        # column mass) would pass the bare C == 0 check, hiding a real
-        # stoichiometry bug elsewhere.
-        assert mass_atm.get('CO2', 0.0) == pytest.approx(0.0, abs=1e-30)
-
-
-# ===========================================================================
-# Invariant 10: monotonicity of S2 Gaillard solubility vs reducing fO2
-# ===========================================================================
-
-
-@pytest.mark.unit
-class TestS2SolubilityMonotonicity:
-    """Gaillard sulfide-saturated solubility increases as fO2 decreases
-    at fixed p_S2 and T (the ``+0.5 ln(p_S2/fO2)`` term carries the
-    redox dependence directly). Tested at the solubility-function level
-    to isolate the law from the multi-species solver feedback loops."""
-
-    @pytest.mark.parametrize('p_S2_bar', [0.01, 0.1, 1.0])
-    @pytest.mark.parametrize('T', [1500.0, 1800.0, 2200.0])
-    def test_gaillard_strictly_decreasing_with_oxidation(self, T, p_S2_bar):
-        S2 = SolubilityS2('gaillard')
-        dIWs = [-4.0, -2.0, 0.0, +2.0, +4.0]
-        values = [S2.gaillard(p_S2_bar, T, dIW) for dIW in dIWs]
-        for i in range(len(values) - 1):
-            assert values[i] > values[i + 1], (
-                f'Gaillard ppmw at dIW={dIWs[i]} ({values[i]:.4e}) '
-                f'not greater than at dIW={dIWs[i+1]} ({values[i+1]:.4e}) '
-                f'(T={T}, p_S2={p_S2_bar})'
-            )
-
-        # Discrimination guard: the monotonicity span across the 8-dex
-        # dIW range should be meaningful (the +0.5 ln(p_S2/fO2) term gives
-        # at least a 10x change in ppmw across dIWs in [-4, +4]). A near-
-        # constant function would pass the > check trivially at each step.
-        assert values[0] / values[-1] > 10.0, (
-            f'Gaillard span too small: values[-4]/values[+4] = '
-            f'{values[0] / values[-1]:.2f}, expected > 10x'
-        )
-
-    def test_negative_pressure_returns_zero(self):
-        """Sad-path: at p_S2 < 1e-20 bar the implementation returns 0
-        to avoid log(0). Verify the floor behaves correctly."""
-        S2 = SolubilityS2('gaillard')
-        assert S2.gaillard(1e-30, 1800.0, 0.0) == pytest.approx(0.0, abs=1e-30)
-
-        # Discrimination guard: a stub that always returned 0 would pass.
-        # Confirm a normal p_S2 of 0.1 bar gives a nonzero, finite value
-        # so the floor branch is genuinely distinct from the main path.
-        normal = S2.gaillard(0.1, 1800.0, 0.0)
-        assert normal > 0.0
-        assert math.isfinite(normal)
-
-
-# ===========================================================================
-# Invariant 11: monotonicity of N2 Dasgupta solubility vs reducing fO2
-# ===========================================================================
-
-
-@pytest.mark.unit
-class TestN2SolubilityMonotonicity:
-    """Dasgupta N2 solubility increases as fO2 decreases at fixed
-    p_N2, p_tot, T (the ``-1.6 dIW`` term in the reduced-N branch
-    drives this). Tested at the solubility-function level."""
-
-    @pytest.mark.parametrize('p_N2_bar', [0.1, 1.0, 10.0])
-    @pytest.mark.parametrize('T', [1500.0, 1800.0, 2200.0])
-    def test_dasgupta_monotonic_with_oxidation(self, T, p_N2_bar):
-        N2 = SolubilityN2('dasgupta')
-        dIWs = [-6.0, -4.0, -2.0, 0.0, +2.0, +4.0]
-        values = [N2.dasgupta(p_N2_bar, p_N2_bar, T, dIW) for dIW in dIWs]
-        # Dasgupta has two terms (reduced-N and molecular N2); the
-        # reduced-N term dominates at low fO2 and falls steeply with
-        # increasing fO2, so the total is monotonically decreasing.
-        for i in range(len(values) - 1):
-            assert values[i] >= values[i + 1], (
-                f'Dasgupta ppmw at dIW={dIWs[i]} ({values[i]:.4e}) '
-                f'less than at dIW={dIWs[i+1]} ({values[i+1]:.4e}) '
-                f'(T={T}, p_N2={p_N2_bar})'
-            )
-
-        # Discrimination guard: the reduced-N branch carries an exp(-1.6 dIW)
-        # factor, so the span across dIW in [-6, +4] should be at least
-        # exp(1.6 * 10) ~ 9e6 in the reduced-N contribution. The molecular
-        # N2 term puts a floor under the high-dIW end, but the span should
-        # still be > 10x. A near-constant function would pass the >= check.
-        assert values[0] / values[-1] > 10.0, (
-            f'Dasgupta span too small: values[-6]/values[+4] = '
-            f'{values[0] / values[-1]:.2f}, expected > 10x'
-        )
-
-    def test_libourel_alternative_is_redox_independent(self):
-        """Sad-path / contrast: the Libourel linear Henry's law has no
-        fO2 term, so it does NOT show the Dasgupta monotonicity. This
-        documents that ``libourel`` and ``dasgupta`` are distinct laws
-        with distinct calibration footprints."""
-        N2 = SolubilityN2('libourel')
-        # Libourel takes only p_N2; no fO2 dependence
-        assert N2.libourel(1.0) == N2.libourel(1.0)  # trivially deterministic
-        # Two different p_N2 give different values (linear)
-        assert N2.libourel(2.0) == pytest.approx(2.0 * N2.libourel(1.0))
-
-
-# ===========================================================================
 # Tim-flagged addition: Dasgupta plateau behaviour at strongly-reducing fO2
 # ===========================================================================
 
 
-@pytest.mark.smoke
 class TestDasguptaReducingEdgeSolver:
     """Buffered mode at the reducing edge of the Dasgupta N
     calibration footprint. The paper Fig. 7 notes plateau-like N
@@ -599,29 +448,11 @@ class TestDasguptaReducingEdgeSolver:
         assert result['N_kg_liquid'] >= 0.0
 
 
-@pytest.mark.unit
-class TestDasguptaReducingEdgeLaw:
-    """Direct check that the Dasgupta law itself stays numerically
-    well-behaved at strongly-reducing conditions, independent of the
-    multi-species solver."""
-
-    @pytest.mark.parametrize('dIW', [-3.0, -4.0, -6.0, -8.0])
-    def test_dasgupta_law_finite_at_reducing_edge(self, dIW):
-        """The reduced-N branch grows as exp(-1.6 dIW); at dIW=-8 the
-        value reaches 10^5 ppmw at p_N2=1 bar but is still
-        numerically representable."""
-        N2 = SolubilityN2('dasgupta')
-        val = N2.dasgupta(1.0, 1.0, 1800.0, dIW)
-        assert math.isfinite(val)
-        assert val > 0.0
-
-
 # ===========================================================================
 # Tim-flagged addition: S2 / SO2 branch at the oxidising calibration edge
 # ===========================================================================
 
 
-@pytest.mark.smoke
 class TestGaillardOxidisingEdge:
     """Gaillard sulfide-saturated calibration ends near IW+3.5
     (FMQ+0.1). Above that the sulfate regime begins and CALLIOPE's
@@ -666,7 +497,6 @@ class TestGaillardOxidisingEdge:
 # ===========================================================================
 
 
-@pytest.mark.smoke
 class TestPGuessWarmStart:
     """A warm start with the cold-solve result lands in the same
     basin and converges to the same partial pressures. PROTEUS
